+++ tests/api_mocks.py
@@ -79,6 +79,7 @@ def _run(name='test'):
         'name': name,
         'state': "running",
         'config': '{"epochs": {"value": 10}}',
+        'description': "",
         'systemMetrics': '{"cpu": 100}',
         'summaryMetrics': '{"acc": 100, "loss": 0}',
         'history': [
@@ -90,7 +91,8 @@ def _run(name='test'):
             '{"cpu": 10}',
             '{"cpu": 20}',
             '{"cpu": 30}'
-        ]
+        ],
+        'tags': []
     }