+++ wandb/wandb_run.py
@@ -6,6 +6,7 @@ import json
 import yaml
 from sentry_sdk import configure_scope
 
+from . import env
 import wandb
 from wandb import history
 from wandb import jsonlfile
@@ -32,7 +33,9 @@ DESCRIPTION_FNAME = 'description.md'
 
 
 class Run(object):
-    def __init__(self, run_id=None, mode=None, dir=None, group=None, job_type=None, config=None, sweep_id=None, storage_id=None, description=None, resume=None, program=None, wandb_dir=None):
+    def __init__(self, run_id=None, mode=None, dir=None, group=None, job_type=None,
+                 config=None, sweep_id=None, storage_id=None, description=None, resume=None,
+                 program=None, wandb_dir=None, tags=[]):
         # self.id is actually stored in the "name" attribute in GQL
         self.id = run_id if run_id else util.generate_id()
         self.resume = resume if resume else 'never'
@@ -81,6 +84,7 @@ class Run(object):
         # important that we overwrite empty strings here.
         if not self.description:
             self.description = self.id
+        self.tags = tags
 
         self.sweep_id = sweep_id
 
@@ -129,10 +133,10 @@ class Run(object):
         """
         if environment is None:
             environment = os.environ
-        run_id = environment.get('WANDB_RUN_ID')
-        resume = environment.get('WANDB_RESUME')
-        storage_id = environment.get('WANDB_RUN_STORAGE_ID')
-        mode = environment.get('WANDB_MODE')
+        run_id = environment.get(env.RUN_ID)
+        resume = environment.get(env.RESUME)
+        storage_id = environment.get(env.RUN_STORAGE_ID)
+        mode = environment.get(env.MODE)
         disabled = InternalApi().disabled()
         if not mode and disabled:
             mode = "dryrun"
@@ -143,17 +147,18 @@ class Run(object):
             wandb.termlog(
                 'W&B is disabled in this directory.  Run `wandb on` to enable cloud syncing.')
 
-        group = environment.get('WANDB_RUN_GROUP')
-        job_type = environment.get('WANDB_JOB_TYPE')
-        run_dir = environment.get('WANDB_RUN_DIR')
-        sweep_id = environment.get('WANDB_SWEEP_ID')
-        program = environment.get('WANDB_PROGRAM')
-        wandb_dir = environment.get('WANDB_DIR')
+        group = environment.get(env.RUN_GROUP)
+        job_type = environment.get(env.JOB_TYPE)
+        run_dir = environment.get(env.RUN_DIR)
+        sweep_id = environment.get(env.SWEEP_ID)
+        program = environment.get(env.PROGRAM)
+        wandb_dir = env.get_dir()
+        tags = env.get_tags()
         config = Config.from_environment_or_defaults()
         run = cls(run_id, mode, run_dir,
                   group, job_type, config,
                   sweep_id, storage_id, program=program,
-                  wandb_dir=wandb_dir,
+                  wandb_dir=wandb_dir, tags=tags,
                   resume=resume)
         return run
 
@@ -261,7 +266,8 @@ class Run(object):
         upsert_result = api.upsert_run(id=id or self.storage_id, name=self.id, commit=api.git.last_commit,
                                        project=project, entity=api.settings(
                                            "entity"),
-                                       group=self.group,
+                                       group=self.group, tags=self.tags if len(
+                                           self.tags) > 0 else None,
                                        config=self.config.as_dict(), description=self.description, host=socket.gethostname(),
                                        program_path=program or self.program, repo=api.git.remote_url, sweep_name=self.sweep_id,
                                        summary_metrics=summary_metrics, job_type=self.job_type, num_retries=num_retries)
@@ -274,23 +280,25 @@ class Run(object):
         """
         if environment is None:
             environment = os.environ
-        environment['WANDB_RUN_ID'] = self.id
-        environment['WANDB_RESUME'] = self.resume
+        environment[env.RUN_ID] = self.id
+        environment[env.RESUME] = self.resume
         if self.storage_id:
-            environment['WANDB_RUN_STORAGE_ID'] = self.storage_id
-        environment['WANDB_MODE'] = self.mode
-        environment['WANDB_RUN_DIR'] = self.dir
+            environment[env.RUN_STORAGE_ID] = self.storage_id
+        environment[env.MODE] = self.mode
+        environment[env.RUN_DIR] = self.dir
 
         if self.group:
-            environment['WANDB_RUN_GROUP'] = self.group
+            environment[env.RUN_GROUP] = self.group
         if self.job_type:
-            environment['WANDB_JOB_TYPE'] = self.job_type
+            environment[env.JOB_TYPE] = self.job_type
         if self.wandb_dir:
-            environment['WANDB_DIR'] = self.wandb_dir
+            environment[env.DIR] = self.wandb_dir
         if self.sweep_id is not None:
-            environment['WANDB_SWEEP_ID'] = self.sweep_id
+            environment[env.SWEEP_ID] = self.sweep_id
         if self.program is not None:
-            environment['WANDB_PROGRAM'] = self.program
+            environment[env.PROGRAM] = self.program
+        if len(self.tags) > 0:
+            environment[env.TAGS] = ",".join(self.tags)
 
     def _mkdir(self):
         util.mkdir_exists_ok(self._dir)